Thunks

Documentation for Thunks.

Thunks.Thunk

Thunks._find_symbols_in_ast

Thunks._thunkify_expr

Thunks.find_assigned_symbols_in_ast

Thunks.find_symbols_in_ast

Thunks.maybe_thunk

Thunks.reify

Thunks.thunkify

Thunks.@lazy

Thunks._find_symbols_in_ast — Method

Return either (nothing, List[Expr]), (Symbol, nothing), or (nothing, nothing).

Helper function following cons and nil pattern from Lisp.

Thunks._thunkify_expr — Method

Wrap expression in anonymous function, with each assigned symbol as an arg.

Thunks.find_assigned_symbols_in_ast — Method

We find all "assigned symbols", as in, everything except:

anonymous functions :((x,y,z) -> x*x).args[1]

generators :((x for x in 1:10))

... others that we should catch...?

Thunks.find_symbols_in_ast — Method

Walk an AST and find all unique symbols.

Thunks.maybe_thunk — Method

Return true if symbol could be assigned to a Thunk.

Thunks.reify — Method

reify(thunk::Thunk)
reify(value::Any)

Reify a thunk into a value.

In other words, compute the value of the expression.

We walk through the thunk's arguments and keywords, recursively evaluating each one, and then evaluating the thunk's function with the evaluated arguments.

Thunks.thunkify — Method

Turn expression into a thunk. Not intended for public usage.

Thunks.Thunk — Type

Thunk(function, args, kwargs)

Type that represents a thunk. Only evaluated is a public field. Use reify to evaluate.

Thunks.@lazy — Macro

@lazy <Expr>

Lazy evaluation of arbitrary expressions.

@lazy x+y
@lazy x = f(y)
@lazy begin
    a = 1
    b = 2
    c = 3
    abc = sum([a,b,c])
end