Thunks

Documentation for Thunks.

Thunks.Thunk

Thunks.reify

Thunks.thunkify

Thunks.@thunk

Thunks.Thunk — Type

Thunk(function, args, kwargs)

Type that represents a thunk. Useful fields inclue:

    evaluated::Bool
    result::Any

Thunks.reify — Method

reify(thunk::Thunk)
reify(value::Any)

Reify a thunk into a value.

In other words, compute the value of the expression.

We walk through the thunk's arguments and keywords, recursively evaluating each one, and then evaluating the thunk's function with the evaluated arguments.

Thunks.thunkify — Method

Turn expression into a thunk. Supports :call, :(=), :block.

Not intended for public usage.

Thunks.@thunk — Macro

@thunk

Macro for turning an expression into a thunk. Supports lines like:

@thunk x+y
@thunk x = f(y)
abc = @thunk begin
    a = 1
    b = 2
    c = 3
    sum([a,b,c])
end