Thunks

Documentation for Thunks.

Thunks.Think

Thunks.Thunk

Thunks.Unevaluated

Thunks._find_symbols_in_ast

Thunks._thunkify

Thunks._thunkify_eq

Thunks._thunkify_expr

Thunks.find_assigned_symbols_in_ast

Thunks.find_symbols_in_ast

Thunks.maybe_thunk

Thunks.reify

Thunks.thunkify

Thunks.thunkify_block

Thunks.@lazy

Thunks.@noeval

Thunks._find_symbols_in_ast — Method

Return either (nothing, List[Expr]), (Symbol, nothing), or (nothing, nothing).

Helper function following cons and nil pattern from Lisp.

Thunks._thunkify — Function

Helper function to dispatch based on Expr head.

Thunks._thunkify_eq — Function

Assign left of = to a thunk.

Thunks._thunkify_expr — Function

Actually create the Thunk struct.

Wrap expression in anonymous function, with each assigned symbol as an arg.

Thunks.find_assigned_symbols_in_ast — Method

We find all "assigned symbols", as in, everything except:

anonymous functions :((x,y,z) -> x*x).args[1]

generators :((x for x in 1:10))

Reserved keywords

... others that we should catch...?

Thunks.find_symbols_in_ast — Method

Walk an AST and find all unique symbols.

Thunks.maybe_thunk — Method

Return true if symbol could be assigned to a Thunk.

Thunks.reify — Method

reify(thunk::Think)
reify(value::Any)

Reify a thunk into a value.

In other words, compute the value of the expression.

We walk through the thunk's arguments and keywords, recursively evaluating each one, and then evaluating the thunk's function with the evaluated arguments.

Thunks.thunkify — Function

Rewrite expression for lazy evaluation using thunks. Not intended for public usage.

Thunks.thunkify_block — Function

Rewrite a begin...end block to be lazy.

Thunks.Think — Type

Abstract type for Thunks.

Thunks.Thunk — Type

Thunk(function, args, kwargs)

Type that represents a thunk. Only evaluated is a public field. Use reify to evaluate. Evaluates to Weak head normal form, meaning that non-free symbols must be assigned / defined. Contrast with Unevaluated.

Thunks.Unevaluated — Type

Keep expression, including args/kwargs, as unevaluated.

Thunks.@lazy — Macro

@lazy <Expr>

Lazy evaluation of arbitrary expressions.

@lazy x+y
@lazy x = f(y)
@lazy begin
    a = 1
    b = 2
    c = 3
    abc = sum([a,b,c])
end
reify(abc)
# output
6

Thunks.@noeval — Macro

@noeval <Expr>

Lazy evaluation of arbitrary expressions that does not check for variable definition. This is safe to use if the assumption of referential transprency holds, and should be used with extreme caution otherwise, e.g. if referenced variables might be mutated. Note that evaluation occurs in the scope of where reify is called.

@noeval begin
    # will not error
    not_used = not_defined + 1
    a = 1
    b = 2
    c = 3
    abc = sum([a,b,c])
end
c = -1
reify(abc)
# output
2